add python stuf
doesn't really work though. `ruff` is super anemic and `pyright` is not
working it crashes with nothing to say in lsp.log  :shrug:

i'll need to keep track of this. I'm guessing it has something to do
with nodejs environment from what I saw on reddit and github but I'm
running latest node (24) and then used `pyright[nodejs]` to use the (22)
version that pyright wants but still just crashes